requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
playwright==1.40.0
selenium==4.15.2
pandas==2.1.3
boto3==1.34.0
asyncio==3.4.3
aiofiles==23.2.1
httpx==0.25.2
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
loguru==0.7.2
schedule==1.2.0 